==> 1_moldesc.ipynb <==
import numpy as np
import pandas as pd
from rdkit import Chem
from rdkit.Chem.SaltRemover import SaltRemover
from mordred import Calculator, descriptors

# === Set project ===
proj = 'LSI'
proj = 'RS'

# === Read in files with X,y ===
Xy = pd.read_csv(proj+'_dat.csv')
clist = list(Xy['SMILES'])
y = np.array(Xy['Label'])
y = np.reshape(y,(-1,1))
X = np.zeros(shape=(len(clist),1826))

# === Calculate descriptors ===
calc = Calculator(descriptors)
dlist = list(calc._name_dict.keys())
remover = SaltRemover()
for i in range(len(clist)):
    mol = Chem.MolFromSmiles(clist[i])
    X[i,:] = calc(mol)
sh = np.shape(X)
print(f'Shape | raw: {sh}')

# === Write to file ===
X_pd = pd.DataFrame(X, columns=dlist)
X_pd.to_csv('temp_out/'+proj+'_X.csv', header=True)
np.savetxt('temp_out/'+proj+'_y.csv', y, delimiter=',')
